backend/requirements.txt
fastapi
uvicorn
pydantic
openai
langchain
langchain-core
langchain-community
httpx
beautifulsoup4  # For web scraping, if needed
python-dotenv   # For managing API keys securely
black